Fix more warnings; turn off warnings for yajl

+++ app/sheet.c
@@ -660,7 +660,7 @@ const char *display_cell(struct zsvsheet_screen_buffer *buff, size_t data_row, s
   char *str = (char *)zsvsheet_screen_buffer_cell_display(buff, data_row, data_col);
   size_t len = str ? strlen(str) : 0;
   if (len == 0 || has_multibyte_char(str, len < cell_display_width ? len : cell_display_width) == 0)
-    mvprintw(row, col * cell_display_width, "%-*.*s", cell_display_width, cell_display_width - 1, str);
+    mvprintw(row, col * cell_display_width, "%-*.*s", (int)cell_display_width, (int)cell_display_width - 1, str);
   else {
     size_t used_width;
     int err = 0;
